requirements.txt
requests
python-telegram-bot
adafruit-io
newsapi-python